+++ src/components/Graph.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import StyledGraph from '@styles/components/Graph.styled';
+
+const Graph: React.FC = () => <StyledGraph />;
+
+export default Graph;