+++ src/styles/components/Graph.styled.tsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+const StyledGraph = styled.div`
+  width: 500px;
+  height: 250px;
+  outline: 1px solid black;
+`;
+
+export default StyledGraph;